% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton_function.R
\name{skeleton}
\alias{skeleton}
\title{A Skeleton Function}
\usage{
skeleton(dircreate)
}
\arguments{
\item{love}{Do you love being tidy? Defaults to TRUE.}
}
\description{
This function allows you to automatically create an open and repprducable project skeleton that everyone can follow.
}
\examples{
skeleton()
}
\keyword{skeleton}
